Move transport assumptions tests to unit test project
